typo fixes and minor language revisions to readme
to be honest, having tutored students against this assignment, i think it's far too early to be pushing TDD etc and i'm not sure this is the way to do it.
i think people are just getting confused right out the door. for an assignment this early i think it's beneficial to still be more structured and provide more direct guidance (eg have the files there already rather than make the student figure out they even are supposed to do that.)
it's just one too many "what is going on here?" at once for a fresh student. they're not ready to be given this kind of compound challenge. they're already freaking out about having to tackle OOP at all, the extra mysteriousness is not helping.

it's also awkward and unclear what exactly the hints are, they're worded like hints but sometimes they're more like a specification but then sometimes they pull back and try to be hints again.

i think in this case the classic ada formula was not broken. but that's just my thoughts.